Better way to handle dock size "custom" setting
As noted in https://github.com/pop-os/desktop-widget/issues/8, the
previous way this was handled was fairly awkward.

This instead makes the "custom" radio button hidden, so when a custom
size is set, no radio button appears selected. The spin button is always
visible.

I'm not sure if there's a better way to do this, but this should be
better anyway.